Added new test and change FindDErr
Create a new test on the line algorithm and updated the calculation of derived Err to  be able to work with floating points

+++ src/main/scala/BresenhamLineAlgorithm.scala
@@ -14,7 +14,7 @@ object BresenhamLineAlgorithm {
 
   def FindDErr(dx: Int, dy: Int): Double = {
     if(dx != 0)
-      math.abs(dy / dx).toDouble
+      math.abs(((dy).toDouble / dx.toDouble))
     else
       dy
   }
